Move email and phone models to user module
