+++ auth-boundary/sign-in/Authenticator.ts
@@ -1,7 +1,7 @@
 import { isCognitoErrorWithCode } from "../CognitoHelpers"
 import { Auth } from "@aws-amplify/auth"
+import { EmailAddress, USPhoneNumber } from "@user/privacy"
 import { CognitoUser } from "amazon-cognito-identity-js"
-import { EmailAddress, USPhoneNumber } from ".."
 
 export type SignInResult =
   | "success"
