+++ user/privacy/PrivacyFormattable.ts
@@ -1,0 +1,6 @@
+/**
+ * An interface for describing how something can be formatted with respect to privacy.
+ */
+export interface PrivacyFormattable {
+  get formattedForPrivacy(): string
+}
