+++ user/privacy/index.ts
@@ -1,0 +1,3 @@
+export * from "./PrivacyFormattable"
+export * from "./EmailAddress"
+export * from "./USPhoneNumber"
